requirements.txt
Django==1.4.1
argparse==1.2.1
distribute==0.6.24
django-taggit==0.9.3
wsgiref==0.1.2
